+++ example/companies/api_docs.py
@@ -2,7 +2,9 @@ COMPANIES_DATA = {
     "type": "company",
     "id": "12",
     "attributes": {
-        "created": "2018-02-21T14:55:06.734781Z",
+        "created": "2018-03-16T09:38:01.531816Z",
+        "updated": "2018-03-16T09:38:01.531879Z",
+        "reg_code": "287-0513",
         "name": "Turner and Sons",
         "email": "turner@sons.com"
     },
@@ -38,14 +40,16 @@ COMPANIES_LIST_RESPONSE = {
             "type": "company",
             "id": "12",
             "attributes": {
-                "created": "2018-02-21T14:55:06.734781Z",
+                "created": "2018-03-16T09:38:01.531816Z",
+                "updated": "2018-03-16T09:38:01.531879Z",
+                "reg_code": "287-0513",
                 "name": "Turner and Sons",
                 "email": "turner@sons.com"
             },
             "relationships": {
                 "employees": {
                     "meta": {
-                        "count": 3
+                        "count": 2
                     },
                     "data": [
                         {
@@ -69,7 +73,8 @@ COMPANIES_LIST_RESPONSE = {
             "type": "employment",
             "id": "162",
             "attributes": {
-                "created": "2018-02-21T14:55:06.756900Z",
+                "created": "2018-03-16T09:48:11.528352Z",
+                "updated": "2018-03-16T09:48:11.528494Z",
                 "name": "Linda Burgess",
                 "email": "carloswoods@griffin.com",
                 "role": 1
@@ -82,7 +87,8 @@ COMPANIES_LIST_RESPONSE = {
             "type": "employment",
             "id": "91",
             "attributes": {
-                "created": "2018-02-21T14:55:06.755331Z",
+                "created": "2018-03-16T09:48:11.528352Z",
+                "updated": "2018-03-16T09:48:11.528494Z",
                 "name": "Crystal Turner",
                 "email": "collinsheather@mendoza.biz",
                 "role": 1
@@ -98,6 +104,7 @@ COMPANIES_CREATE_REQUEST = {
     "data": {
         "type": "company",
         "attributes": {
+            "reg_code": "123-4567",
             "name": "Turner and Sons",
             "email": "turner@sons.com"
         }
@@ -109,7 +116,8 @@ COMPANIES_CREATE_RESPONSE = {
         "type": "company",
         "id": "12",
         "attributes": {
-            "created": "2018-02-21T14:55:06.734781Z",
+            "created": "2018-03-16T09:38:01.531816Z",
+            "updated": "2018-03-16T09:38:01.531879Z",
             "name": "Turner and Sons",
             "email": "turner@sons.com"
         },
@@ -124,9 +132,9 @@ COMPANIES_CREATE_RESPONSES = [
     (400, {
         "errors": [
             {
-                "detail": "Company with this name already exists.",
+                "detail": "Company with this reg_code already exists.",
                 "source": {
-                    "pointer": "/data/attributes/name"
+                    "pointer": "/data/attributes/reg_code"
                 },
                 "status": "400"
             }
@@ -139,14 +147,16 @@ COMPANIES_READ_RESPONSE = {
         "type": "company",
         "id": "12",
         "attributes": {
-            "created": "2018-02-21T14:55:06.734781Z",
+            "created": "2018-03-16T09:38:01.531816Z",
+            "updated": "2018-03-16T09:38:01.531879Z",
+            "reg_code": "287-0513",
             "name": "Turner and Sons",
             "email": "turner@sons.com"
         },
         "relationships": {
             "employees": {
                 "meta": {
-                    "count": 3
+                    "count": 2
                 },
                 "data": [
                     {
@@ -161,7 +171,7 @@ COMPANIES_READ_RESPONSE = {
             }
         },
         "links": {
-            "self": "%(API_ROOT)s/companies/70/"
+            "self": "%(API_ROOT)s/companies/12/"
         }
     },
     "included": [
@@ -169,7 +179,8 @@ COMPANIES_READ_RESPONSE = {
             "type": "employment",
             "id": "162",
             "attributes": {
-                "created": "2018-02-21T14:55:06.756900Z",
+                "created": "2018-03-16T09:48:11.528352Z",
+                "updated": "2018-03-16T09:48:11.528494Z",
                 "name": "Linda Burgess",
                 "email": "carloswoods@griffin.com",
                 "role": 1
@@ -182,7 +193,8 @@ COMPANIES_READ_RESPONSE = {
             "type": "employment",
             "id": "91",
             "attributes": {
-                "created": "2018-02-21T14:55:06.755331Z",
+                "created": "2018-03-16T09:48:11.528352Z",
+                "updated": "2018-03-16T09:48:11.528494Z",
                 "name": "Crystal Turner",
                 "email": "collinsheather@mendoza.biz",
                 "role": 1
@@ -212,6 +224,100 @@ COMPANIES_DELETE_RESPONSES = [
                 "detail": "This company cannot be deleted",
                 "source": {
                     "pointer": "/data"
+                },
+                "status": "400"
+            }
+        ]
+    }),
+]
+
+
+EMPLOYMENTS_DATA = {
+    "type": "employment",
+    "id": "162",
+    "attributes": {
+        "created": "2018-03-16T09:48:11.528352Z",
+        "updated": "2018-03-16T09:48:11.528494Z",
+        "name": "Linda Burgess",
+        "email": "carloswoods@griffin.com",
+        "role": 1
+    },
+    "relationships": {
+        "company": {
+            "data": {
+                "type": "company",
+                "id": "12"
+            }
+        }
+    },
+    "links": {
+        "self": "%(API_ROOT)s/employments/162/"
+    }
+}
+
+EMPLOYMENTS_CREATE_REQUEST = {
+    "data": {
+        "type": "employment",
+        "attributes": {
+            "email": "carloswoods@griffin.com",
+        },
+        "relationships": {
+            "company": {
+                "data": {"type": "company", "id": "12"}
+            }
+        }
+    }
+}
+
+EMPLOYMENTS_CREATE_RESPONSE = {
+    "data": {
+        "type": "employment",
+        "id": "162",
+        "attributes": {
+            "created": "2018-03-16T09:48:11.528352Z",
+            "updated": "2018-03-16T09:48:11.528494Z",
+            "name": "Linda Burgess",
+            "email": "carloswoods@griffin.com",
+            "role": 1
+        },
+        "relationships": {
+            "company": {
+                "data": {
+                    "type": "company",
+                    "id": "12"
+                }
+            }
+        },
+        "links": {
+            "self": "%(API_ROOT)s/employments/162/"
+        }
+    },
+    "included": [
+        {
+            "type": "company",
+            "id": "12",
+            "attributes": {
+                "created": "2018-03-16T09:38:01.531816Z",
+                "updated": "2018-03-16T09:38:01.531879Z",
+                "reg_code": "287-0513",
+                "name": "Turner and Sons",
+                "email": "turner@sons.com"
+            },
+            "links": {
+                "self": "http://localhost:8330/api/2018-02-21/companies/12"
+            }
+        }
+    ]
+}
+
+EMPLOYMENTS_CREATE_RESPONSES = [
+    (201, EMPLOYMENTS_CREATE_RESPONSE),
+    (400, {
+        "errors": [
+            {
+                "detail": "You are not admin in the specified company",
+                "source": {
+                    "pointer": "/data/attributes/company"
                 },
                 "status": "400"
             }