+++ example/companies/serializers.py
@@ -1,27 +1,25 @@
 from rest_framework_json_api import serializers
 
-from companies.models import Company, Employment
+from tg_apicore.serializers import BaseModelSerializer
+
+from companies.models import Company, Employment, User
 
 
-class EmploymentSummarySerializer(serializers.ModelSerializer):
+class EmploymentSummarySerializer(BaseModelSerializer):
     class Meta:
         model = Employment
-        fields = ['id', 'url', 'created', 'name', 'email', 'role']
+        fields = ['id', 'url', 'created', 'updated', 'name', 'email', 'role']
+        create_only_fields = ['email']
 
-    name = serializers.SerializerMethodField()
-    email = serializers.SerializerMethodField()
-
-    def get_name(self, obj):
-        return obj.user.get_full_name()
-
-    def get_email(self, obj):
-        return obj.user.email
+    name = serializers.CharField(source='user.get_full_name', read_only=True)
+    email = serializers.EmailField(source='user.email')
 
 
-class CompanySummarySerializer(serializers.ModelSerializer):
+class CompanySummarySerializer(BaseModelSerializer):
     class Meta:
         model = Company
-        fields = ['id', 'url', 'created', 'name', 'email']
+        fields = ['id', 'url', 'created', 'updated', 'reg_code', 'name', 'email']
+        create_only_fields = ['reg_code']
 
 
 class EmploymentSerializer(EmploymentSummarySerializer):
@@ -35,10 +33,25 @@ class EmploymentSerializer(EmploymentSummarySerializer):
         'company': CompanySummarySerializer,
     }
 
+    def validate_company(self, value):
+        user = self.context['request'].user
+        if not Employment.objects.filter(company=value, user=user, role=Employment.ROLE_ADMIN).exists():
+            raise serializers.ValidationError("You are not admin in the specified company", code='user_not_admin')
+
+        return value
+
+    def validate(self, attrs):
+        # If user's email was given, use it to look up the actual object (creating it if necessary)
+        if attrs.get('user', {}).get('email'):
+            attrs['user'], _ = User.objects.get_or_create(email=attrs.pop('user')['email'])
+
+        return super().validate(attrs)
+
 
 class CompanySerializer(CompanySummarySerializer):
     class Meta(CompanySummarySerializer.Meta):
         fields = CompanySummarySerializer.Meta.fields + ['employees']
+        read_only_fields = ['employees']
 
     class JSONAPIMeta:
         included_resources = ['employees']