+++ HISTORY.rst
@@ -2,6 +2,16 @@
 History
 =======
 
+Next version
+------------------
+
+* Added PageNotFoundView (JSON-based 404 views)
+* Added DetailSerializerViewSet (different serializers and queryset for list/detail/edit views)
+* Added CreateOnlyFieldsSerializerMixin, ModelValidationSerializerMixin and BaseModelSerializer
+* Renamed APIDocumentationView.get_patterns() to .urlpatterns()
+* Improved example app a lot. It now also includes tests that partially test tg-apicore itself
+
+
 0.1.0 (2018-03-08)
 ------------------